requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
streamlit>=1.28.0
supabase>=2.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
chardet>=5.2.0
openai>=1.0.0
python-multipart>=0.0.6
python-docx>=1.0.0
pypdf>=3.17.0

# 音声処理・話者識別
speechbrain>=0.5.16
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
